+++ src/utils/debug.ts
@@ -1,0 +1,5 @@
+import getDebug from 'debug';
+
+const DEBUG_NAME = 'relative-ci:agent';
+
+export const debug = getDebug(DEBUG_NAME);
