+++ src/utils/get-commit-message.ts
@@ -1,0 +1,13 @@
+import childProcess from 'child_process';
+
+export function getCommitMessage(): string {
+  let message = '';
+
+  try {
+    message = childProcess.execSync('git log -1 --pretty=%B').toString().trim();
+  } catch (error) {
+    console.error('Error reading commit message from git', error);
+  }
+
+  return message;
+}
